+++ core/toolbars/utilities/print/fastprint.py
@@ -40,10 +40,16 @@ class GwFastprint:
         self.rubber_band = tools_gw.create_rubberband(self.canvas)
         self.my_json = {}
 
+        folderpath = tools_gw.get_composers_folderpath()
+        qpt_files = tools_gw.load_qpt_templates_into_project(folderpath)
         composers_list = tools_qgis.get_composer()
         if composers_list == '"{}"':
-            msg = "No composers found."
-            tools_qt.show_info_box(msg, "Info")
+            if folderpath and qpt_files is None:
+                msg = "Your composer's path is bad configured. Please, modify it and try again."
+                tools_qgis.show_warning(msg)
+            else:
+                msg = "No composers found."
+                tools_qt.show_info_box(msg, "Info")
             return
 
         self.initial_rotation = self.canvas.rotation()